feat(ui): app state (#82)
* feat(ui): app state

* mverb: replace manual clamp with .clamp()

* state: rename MixingConsole_T0 to MixingConsoleT0

* ui: remove unused imports

* ui: prefix unused variables with underscore

* ui: remove dead WebSocketConnection struct and unused LibretaktUI fields

* sequencer: derive Default for Step

* ui: collapse nested if let

* ui: rename BUTTON_COLOR to button_color

* main: fix unused import and variables in wasm target

* fix(ci): don't deploy WASM on non-main branches

* feat(ui): follow system theme

+++ src/sample_provider.rs
@@ -58,7 +58,7 @@ impl SampleData {
         let mut sample_data: Vec<f32> = vec![];
 
         while let Some(s) = decoder.next() {
-            sample_data.push(s as f32 / i16::MAX as f32);
+            sample_data.push(s / i16::MAX as f32);
             decoder.next(); // Skip the 2nd channel
         }
